Revert "WIP"
This reverts commit e53bcaac36d2e34170f949611600b537bb937a07.

+++ build_common/android/compatibility/android_cpp11_compat.cpp
@@ -69,7 +69,7 @@ namespace std {
         return os.str();
     }
 
-    std::string to_string(size_t t)
+    std::string to_string(uint32_t t)
     {
         std::ostringstream os;
             os << t;
