+++ build_common/android/compatibility/android_cpp11_compat.h
@@ -16,7 +16,7 @@ namespace std {
 
     std::string to_string(int i);
     std::string to_string(long t);
-    std::string to_string(size_t i);
+    std::string to_string(uint32_t i);
     std::string to_string(double d);
 }
 #endif
