+++ resource/c_common/ocrandom/test/randomtest.cpp
@@ -30,7 +30,7 @@ extern "C" {
 #define ARR_SIZE (20)
 
 TEST(RandomGeneration,OCGetRandom) {
-    size_t value = OCGetRandom();
+    uint32_t value = OCGetRandom();
     EXPECT_LE((uint8_t )0, value);
     EXPECT_GT(pow(2, 32), value);
 }
